v1.0 - First release.
Video and Sound is all working (afaik)
ASM removed/replaced.
Internally it draws everything as if it were still using a CGA screen,
I have written some SDL code that "emulates" a CGA screen and draws
it to an SDL screen.
Sound is done in a similar way: The digital output in LibSDL is used to
make the appropriately-pitched beepings to sound like a PC Speaker :)
The API to the PC Speaker code is identical to those of a small
library I wrote a while back to control the PC Speaker in DJGPP, this
is incase I want to do a DOS port.
Moved all of Andrew Jenners decompiled code into amj/, I do not use this
directly but it is useful in deciphering some ASM functions.
attic/ has all the asm source and other source files that are redundant
now.
doc/ has all the documentation that came with the source
current bugs: crashing into an ox doesnt turn your screen pink,
	map/status bar dont refresh properly (?? score not displayed)