Generalize menu code.
We want to add a second menu for key bindings. Add runmenu() which takes
a list of menuitems to present. As part of this, separate out the
description strings so that they belong to the menu, and are not tied into
the confoption system. This means that the menus can be flexibly designed
and changed without any dependency on the configuration file.