Drop missiles settings from the options menu.
Maybe this will prove controversial but I'm going to do it anyway. I believe
that the missiles feature was a mistake to add to the game - it doesn't fit the
WW1 theme (heat seeking missiles?!), and seriously messes up the gameplay in a
way that really isn't very fun at all, particularly in vs. computer mode where
the enemy planes get heat seekers and you don't. I have a distinct memory of
reading a post by Dave Clark where even he admitted that it was a half-baked
idea, so that's another data point in favor of ditching this.

The code still exists and so do the settings, so if there really are some
die-hard fans who play with missiles enabled, they can edit the config file to
turn them back on. And if anyone strongly disagrees with this change and wants
to argue in favor of this feature then please contact me ...