Turn on medals by default.
Christoph was humble when he implemented this feature and didn't turn it on by
default, but I think it's a great addition to the game - it fits the game's
theme well, doesn't change the original gameplay (getting a medal doesn't do
anything except make one appear on your screen), and is subtle and
non-distracting. So let's turn it on by default because I think more people
deserve to be awarded medals when they do really well in the game.