Fix network game desyncs with harry keys mode.
When playing with harry keys mode we invert the up/down controls to make
turning relative to the screen instead of the plane. However, the code to do
this was assuming that both players had the same conf_harrykeys setting, which
is not necessarily the case. To fix this, include the local value of
conf_harrykeys as a new bit when building a player command, so that it gets
exchanged over the network with the other player.  We then do the inversion
based on that particular player's conf_harrykeys setting, rather than our own
local setting.