Use SDL_GetPrefPath() to get a config file location.
SDL_GetPrefPath()'s documentation says that "You should assume the path
returned by this function is the only safe place to write files", which means
that it's a good place to save our config file. This does mean that the config
file location has changed on *nix, but the path returned by SDL in
~/.local/share is a better place nowadays anyway, rather than cluttering the
homedir. It also means that we're now writing to a known good location on
Windows rather than just dumping a .ini file in the current directory.