Expand and improve the main manpage significantly.
I was originally planning to just add references to the sopwith.cfg
manpage, but got carried away. Things added: location of the configuration
file, a list of gameplay tips, names for the ribbons along with descriptions of
what they look like, and a paragraph about the real Sopwith Camel to the
history section.