Fix library ordering issue with compiles on Windows.
Compilation for mingw was failing with the error "Undefined reference to
SDL_main" because libSDLmain could not find the SDL_main function (our main
function gets renamed to SDL_main by a #define set by libSDL). This appears to
be something like a circular dependency because the ordering of libraries when
linking affects symbol resolution. To fix the issue, split out main.c from
libsdlsopwith.a into libsdlsopmain.a - this can go before @SDL_LIBS@ while
libsdlsopwith.a goes after it.