Rework how we set the application icon.
Now that we're on SDL2 we can set much larger icon sizes, and this lets
us avoid getting a blurry upscaled icon on eg. macOS. Previously the video
code set a random plane icon but some of the rotated plane sprites are hard
to make out when seen in isolation. So instead, set the normal starting
sprite as it's more distinctive, but with a little easter egg thrown in.