Fix up big oil tank explosions.
Jesse tweaked these in a previous version but they were still nowhere
near as intense as they are in the original Sopwith 1/Sopwith 2. I got
the +4 value here from looking at Andrew Jenner's disassembly, and it
seems to match the DOS versions. As part of this, replace the command
line argument with a configuration variable for consistency.