Fix fractional movement calculations.
The `movexy` and `setdxdy` functions have been incomplete since the
very first 1.0 release of SDL Sopwith in that they were not taking
fractional position and momentum values into account, only integer
ones. This means that all movement within the game was significantly
less precise than it ought to have been.

What led to this discovery was a bug where it was discovered that it
was possible to stop in mid-air in novice mode. The game code still
considered the plane to be moving, but without fractional momentum
values being included in calculations, the plane would not move.
Normally the plane stalls at low speeds, masking the bug. It was only
in novice mode where it was impossible to stall that the bug surfaced.
Thanks to Alaux for reporting this bug.

Fixes #3.