Fix a crash where birds can get stuck inside the floor.
The previous change that fixed the movement code seems to have made
this start occurring and it is possible the bug was being masked by
the other bug. This adds a simple check to ensure that birds bounce
off the floor in a way that makes them fly away from walls.
Anything ground immediately to the left/right that is higher than
roughly a 30 degree slope counts as a wall. If the new direction
would have us fly towards a wall, we go the other way instead.