Fix x-position linked list.
The list insert/delete code had some assumptions baked in which are not
necessarily true - that we won't be inserting the object at either the top or
bottom of the list - and are part of why the previously fixed bug with birds
getting stuck caused the game to crash. So let's take the opportunity to fix
this and handle the corner cases correctly.