Add filter for PC speaker sound output.
Product of this weekend's hacking. This adds a bandpass filter that
attempts to simulate how a PC speaker really sounds when playing
Sopwith on an original PC. We filter out the low end (since the PC
speaker is a tiny speaker without any kind of bass response), and the
high end as well (since the PC speaker is mounted inside a metal box
that muffles any output). The frequency range of 1500-6000 Hz is based
on frequency analysis of an LGR video where he plays Silpheed on an
original IBM 5150 PC.