Add config option to disable tinny filter.
Maybe there's someone out there who really likes the unfiltered
output, so let's include a config option. But I have sufficient faith
in the decision to add the filter that I'm leaving this turned on by
default and keeping it as a config file-only option without bothering
to include it in the menu.