Eliminate use of MAX_X in MAXFUEL constant.
The maximum fuel was being defined in terms of the level size, which
makes some sense. However, I think that for custom levels this is likely
to prove more confusing than sensible. It makes more sense if the
maximum fuel remains the same and it may even prove useful as a game
mechanic for certain levels (eg. to travel a long way and back without
running out of fuel).