Fix ox splat to display if we are already shot down.
In the case where our plane is shot down (or we hit a bird) and we then
crash into an ox, the bullet and bird splats accumulate on the screen as
we take more damage but the final screen inversion was not occurring
when we hit the ox. This appears to be due to some duplicated code where
the ox check was not included. Let's fix this for consistency.