Fix bug where you can win by bombing your own buildings.
This bug was mistakenly introduced in commit bfd67c when I changed the
way that the numtarg[] array gets initialized. I did not realise that a
value of zero had a special meaning where that target type gets ignored;
in other words, the win condition was being set at level initialization
time, not at the time when a target was destroyed. I have changed the
check to the latter which I believe is more expected.