Use x-position linked list for autopilot AI.
Since we already have this linked list of all the in-game objects,
it's not necessary to use a separate array of all the target objects
for the AI. The change here is careful to preserve the previous logic,
and I've tested by playing the game with both versions of the logic in
parallel to confirm they give the same result, resolving any detected
discrepancies. However, the logic that calculates the X range of objects
to test is probably unnecessarily complex.