Add error_exit() function to exit with message.
The new function prints an error to stderr if running in a terminal, but
if not, uses the SDL API to show an error message. This should be more
friendly on systems like Windows where it won't usually be invoked from
the terminal, but also on Linux where users might launch the game from
the desktop and not see any error message otherwise.