Don't allow planes to get wounded on the ground.
I noticed a bug where your plane could get wounded sitting on the runway
and still be damaged after takeoff without being repaired. Rather than fix
the repair logic, I've decided that wounding just shouldn't happen on
players sitting on the ground. It ought to be possible to "raid" planes that
are sitting on the ground without the damage being repaired instantly.