More tweaks to the beginner's help screen.
First of all, we should only ever show help on the first level.

Secondly, make the help screen appear less often in novice mode. It's not
necessary to keep showing it after the player has demonstrated that they know
the controls.  We identify this by tracking how long the player manages to keep
the plane in the air; if they manage to successfully fly it for at least 8
seconds without crashing, that's considered a successful flight and we won't
show help again.

Third, show the help screen as a hint in the other single player modes as well:
but only if they appear to be really struggling to get the plane off the
ground. We identify this if they crash the plane three times without managing a
single successful flight. This seems like a reasonably conservative metric that
shouldn't annoy experienced players by putting a bunch of text on the screen
that they won't care about.

To support the latter, also move the help text onto the left hand side of the
screen. This is because the right side is the side from which the enemy plane
will flying from to attack the player in "vs. computer" mode. We don't want the
text to obscure this.