Get rid of oobjects[] array.
We only have this to save the original positions of planes so that they can
find their way back to base later. This is duplicating (some of) the data
we already have saved in the original_ob_t structure. The only exception is
the original Y position, so let's just add (yet) another field to the
object struct to save it.

As an alternative I considered putting the new field in original_ob_t, but
this would mutate the original level definition which I want to keep
constant.