Constify game and original_ob_t pointers.
Once a level has been loaded, it should never be modified by the game
code. This also makes the original_level struct const, although note
that the ground and object array pointers are not; those fields need
to be mutable so that we can construct them later when loading levels.