Eliminate nobjects[] pre-allocated array.
Objects were being allocated at the start of the level as a single contiguous
array of OBJECTS structures; this meant that we had a static limit of 100
objects that could be instantiated at once. There were also places where eg.
nobjects[0] was being used to refer to player 1's plane because
(coincidentally) that player would be the first object allocated when the level
is created.

This gets rid of the nobjects[] array entirely. For those cases whhere we used
it to number planes, add a new planes[] array. We also get rid of the ob_index
field, since this no longer has any meaning now that the objects are no longer
allocated contiguously.