Eliminate gm_planes[] array, merge with gm_targets[].
The level format is more complicated if we have to have separate object
lists. Let's instead merge the planes and targets list into a single list
for simplicity. In order to be able to look up plane objects by player
number, add an orig_planes[] array that gets generated when the level
starts.