video: Disable text input on startup.
The most recent versions of SDL2 seem to start up the game with text
input enabled. We don't want this; text input should only be enabled
if we're typing something in (= server address for a network game).
Having it enabled like this breaks the key rebindings screen.

Fixes bug reported on the Sopwith Discord by ranguli; thanks also to
NY00123 for investigations.