Alternative fix for invisible birds bug.
This eliminates the use of an invalid owner number by changing the
ob_clr value from 9 to 1. This fixes the birds being invisible and also
correctly renders the birds on the map in cyan. I also like this fix
because we already set the same value for the oxen, and the player 1
"color scheme" is essentially "just use the normal colors".

My one concern with this fix was that changing the color value might
change some game behavior that takes it into account. To guard against
this, I carefully inspected all the game code that references the
`ob_clr` field, and I'm satisfied that for flocks and birds the `ob_clr`
field isn't used for anything else apart from when drawing the screen.