Use color mapping for map pixel color.
This was one aspect of @techknight's fix in 8ffae0862efd171fc that I
really liked. Rather than just using the color mapping number modulo 4,
let's be more deliberate and look up an actual color that corresponds
to the mapping. In the case of the planes this is the fuselage color.