Patch configure.ac version before CI build.
This gives a more meaningful version number in the CI builds instead
of the version number of the last release. It's unfortunate that the
only apparent way to do this is by patching the file, because autoconf
requires the version number given to AC_INIT to be a string literal.

Thanks to @NY00123 for the suggestion on Discord.