Change semantics of updateobjpos() function.
The name is misleading because it doesn't suggest that the object would
be implicitly added into the X list. Let's reduce the scope of the
function to only ever update the position of objects that are already
in the list; insertx() keeps its old job of inserting into the list.

This rolls back the previous change to swinit.c that was made in
commit b35df6e36f9b88ab24441d646d6f631070bdabce.