Call updateobjpos() implicitly from movexy().
This allows us to eliminate almost all of the direct calls to
updateobjpos(), which helps to further make the interface more
idiot-proof.

As part of this, we have to handle the fact that the autopilot code does
speculative movements on copies of the real objects. So we add a
copyobj() function that resets the linked list pointers, so that
movexy() does not try to hook the copied objects back into the X list.