Fix bug with planes stalling immediately on takeoff.
Enemy planes were exhibiting this behavior especially if the player
was behind them or directly overhead. They would pull the stick back
and try to fly straight upwards without being at full acceleration.
The cause appears to be that since any low altitude was considered
a crash, they would not accelerate and could not choose a sensible
direction to aim for. Making the low altitude check only apply when
aiming downwards fixes the problem.

This change might be considered controversial since it changes the
behavior of the AI. I did bear this in mind when making the change,
but decided to go ahead and make it, because:

 * Sopwith 2 does not actually appear to have this bug; it first
   appears in the network edition. So this is not some aspect of
   the AI that can be argued is a fundamental part of the game.
 * It's a really embarrassingly silly bug that deserves to be fixed.