Add Atari palette.
This is defined in swgrpha.c in the original sources and appears to
be the palette that would have been used in the Atari port. The
red/blue color scheme also conveniently matches the terminology used
in the original documentation where the planes are referred to as being
red and blue.