autopilot: Use predicted y speed, not current
To avoid crashing into the ground we try to avoid aiming downwards if
we're flying low. However, the downwards check (`dy < 0`) was being
performed against the *current* downwards momentum and not the predicted
momentum if a particular direction is chosen. This could lead the
autopilot to aim downwards if they were just above the low-flying
threshold, causing a crash if the plane is flying at high speed.

Part of an attempted fix for #44.