Build up linked list of all symbols
To be able to replace symbols, we need to be able to find them. So when
initializing symbols, build up a linked list of all the symbols in the
game. Later, the lookup function will be able to scan through the linked
list to find particular symbols by name and frame number.