sound: Change music parser to use C major scale
The music code interprets a format that assumes each octave begins with
'A', which - although a reasonable assumption - is not the case in the
C major scale that is most common in western music. Since we're adding
support for user-defined tunes, it seems like a good idea to switch to
the more conventional scale to avoid confusion.