Simplify a lot of the plane state checks
A lot of the object states are pseudo-duplicates - it is not possible to
just check if the state is `FLYING` for example, because it might also be
`WOUNDED`, `STALLED` or `WOUNDSTALL`. To simplify some complicated `if`
statements that check plane states, this adds convenience functions that
make the intent clearer.