Add ten user-defined target types
These are generic target types that the map author can use to create
their own custom targets. The base symbol is just a plain rectangle; the
expectation is that the `symbol` block will be used to replace it. I've
extended the `new_targets.sop` map file to use one of the slots as an
example.

Since we have a lot of target types now, I've also defined an enum to
make referring to them clearer.