Combine `ob_dy` and `ob_ldy` fields
In the original Sopwith these were both 16-bit integers, with the
`ob_ldy` field containing the fractional part of the Y momentum for
precision. However, the resulting math is overly complicated because the
momentum is split across two fields. Nowadays we can just represent both
as a 32-bit integer.

The replacement field is named `ob_ndy` ("new" dy) to distinguish it.